requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2
django-appconf==1.0.5
django-environ==0.10.0
django-staticfiles==1.2.1
django-webpack-loader==1.8.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==1.0.0
python-decouple==3.8
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
urllib3==2.0.2
whitenoise==6.4.0